storage-api/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic for data validation
pydantic>=2.8.0
pydantic-settings>=2.2.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Authentication
PyJWT==2.8.0
python-multipart==0.0.6

# File handling and validation
python-magic==0.4.27

# HTTP client
httpx==0.25.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production dependencies
gunicorn==21.2.0
